Add Fixed Size Sample to ml_metrics.
This metric is used to generate a fixed size sample of the data stream. Initially, the reservoir of datastream values will grow to the size of 'size'. Then, subsequent values may replace values in the reservoir, depending on random sampling.

This is also known as Reservoir Sampling: https://en.wikipedia.org/wiki/Reservoir_sampling.

PiperOrigin-RevId: 681154441